requirements.txt
PySide6>=6.6.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
python-binance>=1.0.19
pytest>=8.0.0
